NETRA_TAX/backend/requirements.txt
# NETRA TAX Backend Requirements
# Python 3.9+

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.1
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
torch==2.1.1
torch-geometric==2.4.0
scikit-learn==1.3.2
networkx==3.2

# Database (PostgreSQL)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# PDF Generation
reportlab==4.0.7
pypdf==3.17.1
Pillow==10.1.0

# Email & Notifications
python-emailfastapi==0.0.3
aiosmtplib==3.0.1

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.0

# System Utilities
psutil==5.9.6
requests==2.31.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
